refactor(store): add shared database helpers
